+++ crest/Assets/Crest/Crest/Scripts/Underwater/UnderwaterRenderer.Effect.cs
@@ -37,6 +37,7 @@ namespace Crest
 
         CommandBuffer _underwaterEffectCommandBuffer;
         PropertyWrapperMaterial _underwaterEffectMaterial;
+        Material _currentOceanMaterial;
         internal readonly UnderwaterSphericalHarmonicsData _sphericalHarmonicsData = new UnderwaterSphericalHarmonicsData();
 
         RenderTargetIdentifier _colorTarget = new RenderTargetIdentifier
@@ -243,7 +244,7 @@ namespace Crest
             }
         }
 
-        internal static void UpdatePostProcessMaterial(
+        internal void UpdatePostProcessMaterial(
             Mode mode,
             Camera camera,
             PropertyWrapperMaterial underwaterPostProcessMaterialWrapper,
@@ -256,10 +257,37 @@ namespace Crest
         )
         {
             Material underwaterPostProcessMaterial = underwaterPostProcessMaterialWrapper.material;
-            if (copyParamsFromOceanMaterial)
+
+            // Copy ocean material parameters to underwater material.
             {
-                // Measured this at approx 0.05ms on dell laptop
-                underwaterPostProcessMaterial.CopyPropertiesFromMaterial(OceanRenderer.Instance.OceanMaterial);
+                var material = OceanRenderer.Instance.OceanMaterial;
+                // Grab material from a water body if camera is within its XZ bounds.
+                foreach (var body in WaterBody.WaterBodies)
+                {
+                    if (body._overrideMaterial == null)
+                    {
+                        continue;
+                    }
+
+                    var bounds = body.AABB;
+                    var position = camera.transform.position;
+                    var contained =
+                        position.x >= bounds.min.x && position.x <= bounds.max.x &&
+                        position.z >= bounds.min.z && position.z <= bounds.max.z;
+                    if (contained)
+                    {
+                        material = body._overrideMaterial;
+                        // Water bodies should not overlap so grab the first one.
+                        break;
+                    }
+                }
+
+                if (copyParamsFromOceanMaterial || material != _currentOceanMaterial)
+                {
+                    // Measured this at approx 0.05ms on Dell laptop.
+                    underwaterPostProcessMaterial.CopyPropertiesFromMaterial(material);
+                    _currentOceanMaterial = material;
+                }
             }
 
             underwaterPostProcessMaterial.SetVector("_DepthFogDensity", OceanRenderer.Instance.UnderwaterDepthFogDensity);